refactor: move some helpers in utils package
