+++ api/kyverno/v1/resource_description_types.go
@@ -3,6 +3,7 @@ package v1
 import (
 	"fmt"
 
+	kubeutils "github.com/kyverno/kyverno/pkg/utils/kube"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/validation/field"
@@ -57,7 +58,7 @@ func (r *ResourceDescription) Validate(path *field.Path, namespaced bool, cluste
 	if r.Name != "" && len(r.Names) > 0 {
 		errs = append(errs, field.Invalid(path, r, "Both name and names can not be specified together"))
 	}
-	if r.Selector != nil && !labelSelectorContainsWildcard(r.Selector) {
+	if r.Selector != nil && !kubeutils.LabelSelectorContainsWildcard(r.Selector) {
 		if selector, err := metav1.LabelSelectorAsSelector(r.Selector); err != nil {
 			errs = append(errs, field.Invalid(path.Child("selector"), r.Selector, err.Error()))
 		} else {
