+++ pkg/engine/wildcards/wildcards.go
@@ -3,9 +3,9 @@ package wildcards
 import (
 	"strings"
 
-	commonAnchor "github.com/kyverno/kyverno/pkg/engine/anchor"
-
 	wildcard "github.com/kyverno/go-wildcard"
+	commonAnchor "github.com/kyverno/kyverno/pkg/engine/anchor"
+	stringutils "github.com/kyverno/kyverno/pkg/utils/string"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -21,7 +21,7 @@ func ReplaceInSelector(labelSelector *metav1.LabelSelector, resourceLabels map[s
 func replaceWildcardsInMapKeyValues(patternMap map[string]string, resourceMap map[string]string) map[string]string {
 	result := map[string]string{}
 	for k, v := range patternMap {
-		if hasWildcards(k) || hasWildcards(v) {
+		if stringutils.ContainsWildcard(k) || stringutils.ContainsWildcard(v) {
 			matchK, matchV := expandWildcards(k, v, resourceMap, true, true)
 			result[matchK] = matchV
 		} else {
@@ -30,10 +30,6 @@ func replaceWildcardsInMapKeyValues(patternMap map[string]string, resourceMap ma
 	}
 
 	return result
-}
-
-func hasWildcards(s string) bool {
-	return strings.Contains(s, "*") || strings.Contains(s, "?")
 }
 
 func expandWildcards(k, v string, resourceMap map[string]string, matchValue, replace bool) (key string, val string) {
@@ -58,8 +54,8 @@ func expandWildcards(k, v string, resourceMap map[string]string, matchValue, rep
 // replaceWildCardChars will replace '*' and '?' characters which are not
 // supported by Kubernetes with a '0'.
 func replaceWildCardChars(s string) string {
-	s = strings.Replace(s, "*", "0", -1)
-	s = strings.Replace(s, "?", "0", -1)
+	s = strings.ReplaceAll(s, "*", "0")
+	s = strings.ReplaceAll(s, "?", "0")
 	return s
 }
 
@@ -145,7 +141,7 @@ func getValueAsStringMap(key string, data interface{}) (string, map[string]strin
 func replaceWildcardsInMapKeys(patternData, resourceData map[string]string) map[string]interface{} {
 	results := map[string]interface{}{}
 	for k, v := range patternData {
-		if hasWildcards(k) {
+		if stringutils.ContainsWildcard(k) {
 			anchorFreeKey, anchorPrefix := commonAnchor.RemoveAnchor(k)
 			matchK, _ := expandWildcards(anchorFreeKey, v, resourceData, false, false)
 			if anchorPrefix != "" {
