+++ pkg/utils/string/wildcard.go
@@ -1,0 +1,7 @@
+package string
+
+import "strings"
+
+func ContainsWildcard(v string) bool {
+	return strings.Contains(v, "*") || strings.Contains(v, "?")
+}
